add new hooks
