+++ client/src/hooks/index.js
@@ -1,1 +1,3 @@
 export { default as useSearchParams } from "./useSearchParams";
+export { usePagination } from "./usePagination";
+export { useFetchTopics } from "./useFetchTopics";
